Trying to implement an update
Should I try to hammer out a PUT, or leave it at POST?
Do I make the modification on the client, then replace the entire
database on the server?
Should I make the modification on the server, and simply respond with
the changed element?
Probably the last one, to avoid sending a buttload of data over the
air.